move misplaced modules / functions
