+++ bentoml/server/bento_api_server.py
@@ -25,7 +25,7 @@ from werkzeug.utils import secure_filename
 
 from bentoml import config
 from bentoml.utils.usage_stats import track_server
-from bentoml.utils.trace import trace
+from bentoml.server.trace import trace
 from bentoml.exceptions import BentoMLException
 from bentoml.server.instruments import InstrumentMiddleware
 from bentoml.server.open_api import get_open_api_spec_json
