+++ bentoml/utils/__init__.py
@@ -24,19 +24,11 @@ from urllib.parse import urlparse, uses_netloc, uses_params, uses_relative
 from google.protobuf.json_format import MessageToDict
 from ruamel.yaml import YAML
 
-from bentoml import __version__ as BENTOML_VERSION, _version as version_mod
+from bentoml import __version__ as VERSIONEER_VERSION, _version as version_mod
 from bentoml.proto import status_pb2
 
 _VALID_URLS = set(uses_relative + uses_netloc + uses_params)
 _VALID_URLS.discard("")
-
-
-def detect_free_port(host='localhost'):
-    sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-    sock.bind((host, 0))
-    port = sock.getsockname()[1]
-    sock.close()
-    return port
 
 
 @contextmanager
@@ -92,7 +84,7 @@ def ProtoMessageToDict(protobuf_msg, **kwargs):
 
 def _is_pypi_release():
     is_installed_package = hasattr(version_mod, 'version_json')
-    is_tagged = not BENTOML_VERSION.startswith('0+untagged')
+    is_tagged = not VERSIONEER_VERSION.startswith('0+untagged')
     is_clean = not version_mod.get_versions()['dirty']
     return is_installed_package and is_tagged and is_clean
 
@@ -115,20 +107,6 @@ def status_pb_to_error_code_and_message(pb_status):
     error_code = status_pb2.Status.Code.Name(pb_status.status_code)
     error_message = pb_status.error_message
     return error_code, error_message
-
-
-def is_postgres_db(db_url):
-    try:
-        return urlparse(db_url).schema.startswith('postgresql')
-    except ValueError:
-        return False
-
-
-def is_sqlite_db(db_url):
-    try:
-        return urlparse(db_url).scheme == 'sqlite'
-    except ValueError:
-        return False
 
 
 def cached_property(method):
