+++ bentoml/yatai/client/deployment_api.py
@@ -32,7 +32,7 @@ from bentoml.proto.deployment_pb2 import (
 )
 from bentoml.exceptions import BentoMLException, YataiDeploymentException
 from bentoml.proto import status_pb2
-from bentoml.utils.validator import validate_deployment_pb_schema
+from bentoml.yatai.validator import validate_deployment_pb_schema
 from bentoml.yatai.deployment_utils import (
     deployment_yaml_string_to_pb,
     deployment_dict_to_pb,
