+++ bentoml/yatai/yatai_service_impl.py
@@ -44,7 +44,7 @@ from bentoml.db import init_db
 from bentoml.yatai.status import Status
 from bentoml.proto import status_pb2
 from bentoml.utils import ProtoMessageToDict
-from bentoml.utils.validator import validate_deployment_pb_schema
+from bentoml.yatai.validator import validate_deployment_pb_schema
 from bentoml import __version__ as BENTOML_VERSION
